feat: 完善图表生成功能

+++ youran-generate-core/src/main/java/com/youran/generate/service/chart/MetaChartService.java
@@ -3,14 +3,17 @@ package com.youran.generate.service.chart;
 import com.youran.common.constant.ErrorCode;
 import com.youran.common.exception.BusinessException;
 import com.youran.generate.dao.chart.MetaChartDAO;
+import com.youran.generate.pojo.po.MetaProjectPO;
 import com.youran.generate.pojo.po.chart.MetaChartPO;
 import com.youran.generate.pojo.qo.chart.MetaChartQO;
 import com.youran.generate.pojo.vo.chart.MetaChartListVO;
+import com.youran.generate.service.MetaProjectService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * 【图表】增删改查服务
@@ -23,8 +26,8 @@ public class MetaChartService {
 
     @Autowired
     private MetaChartDAO metaChartDAO;
-
-
+    @Autowired
+    private MetaProjectService metaProjectService;
 
     /**
      * 查询列表
@@ -63,12 +66,35 @@ public class MetaChartService {
     public int delete(Integer... chartIds) {
         int count = 0;
         for (Integer chartId : chartIds) {
+            MetaChartPO po = metaChartDAO.findById(chartId);
+            if (po == null) {
+                continue;
+            }
+            Integer projectId = po.getProjectId();
+            //校验操作人
+            MetaProjectPO project = metaProjectService.getAndCheckProject(projectId);
             count += metaChartDAO.delete(chartId);
+            metaProjectService.updateProject(project);
         }
         return count;
     }
 
 
+    /**
+     * 根据项目id查询所有图表实体
+     * @param projectId
+     * @param cast
+     * @return
+     */
+    public List<MetaChartPO> findByProjectId(Integer projectId, boolean cast) {
+        List<MetaChartPO> list = metaChartDAO.findByProjectId(projectId);
+        if(!cast){
+            return list;
+        }
+        return list.stream()
+            .<MetaChartPO>map(po -> po.castSubType(true))
+            .collect(Collectors.toList());
+    }
 }
